==> quapos-lm/02-feature-extraction/01-wt/01-a-feature-extraction-simpleitk.ipynb <==
## Feature extraction postnatal development - napari_simpleitk_image_processing

The workflow for  extracting features with `napari_simpleitk_image_processing` and its `label_statistics` function was already demonstrated [here]. 

This notebook aims to extract features from all images in the statistics-dataset folder using a for loop.

import apoc
import numpy as np
import os
import pandas as pd
import pyclesperanto_prototype as cle

from napari_simpleitk_image_processing import label_statistics

import sys
sys.path.append("../..")
from quapos_lm import rescale_image, rescale_segmentation, predict_image

# Load classifier
quapos_lm = apoc.ObjectSegmenter(opencl_filename = "../../01-training-and-validation/02-quapos-lm.cl")
quapos_lm.feature_importances()

Now the folder with the images used for statistical analysis will be loaded and a file list defined.

# Define directory for images
images = "../../data/02-data-for-pixel-classifier/statistics-data/"

# Define file list
file_list = os.listdir(images)

# Show file list
print(file_list)

### Extract features

Now a for loop will be computed to extract all the features. Additionally, information from the filename will be used to add the columns `age`, `biol_repl`, and `image_id`.

# Define empty array to store all extracted features
features = []

# Itterate a for loop over the image folder
for i, file_name in enumerate(file_list):
    
    # Load image
    image = cle.imread(images + file_name)
    
    # Predict the image
    prediction = predict_image(image=image, classifier=quapos_lm)
    
    # Rescale the image
    image_rescaled = rescale_image(image=image, voxel_x=0.323, voxel_y=0.323, voxel_z=0.49)
    
    # Rescale the prediction
    prediction_rescaled = rescale_segmentation(segmentation=prediction, voxel_x=0.323, voxel_y=0.323, voxel_z=0.49)
    
    # Extract features with napari_simpleitk_image_processing's label_statistics function
    # Here all possible features are extracted
    features_i = label_statistics(
        intensity_image=image_rescaled,
        label_image=prediction_rescaled,
        size=True,
        intensity=True,
        perimeter=True,
        shape=True,
        position=True,
        moments=True)
    
    # Add information from the filename into corresponding columns
    # Split the filename
    file_name_split = file_name.split("-")
    
    # Add age in a respective column
    age = int(file_name_split[1].replace("p", ""))
    features_i["age"] = pd.Series([age for x in range(len(features_i))])
    
    # Add the biological replicate in a respective column
    biol_repl = file_name_split[-1].split(".")
    biol_repl = biol_repl[0]
    features_i["biol_repl"] = pd.Series([biol_repl for x in range(len(features_i))])
    
    # Add the image_id in a respective column
    features_i["image_id"] = pd.Series([i for x in range(len(features_i))])
    
    # Store measurements of current image in dataframe
    features.append(features_i)
    
# Concatenate dataframe with all measurements
features = pd.concat(features)

# Show the table with the extracted features
features

features.to_csv("../../measurements/wt-postnatal-development/01-a-feature-extraction-simpleitk.csv", index = False)
