==> quapos-lm/02-feature-extraction/01-wt/02-combine-dataframes.ipynb <==
## Combine dataframes

Features were extracted from all images with [napari_simpleitk_image_processing](01-a-feature-extraction-simpleitk.ipynb), and [porespy](01-b-feature-extraction-porespy.ipynb). Now the two dataframes are combined on the index.

import numpy as np
import pandas as pd

# Define path were measurements are stored
path = "../../measurements/wt-postnatal-development/"

# Load simpleitk measurements
measurements_simpleitk = pd.read_csv(path + "01-a-feature-extraction-simpleitk.csv")
measurements_simpleitk

# Load porespy measurements
measurements_porespy = pd.read_csv(path + "01-b-feature-extraction-porespy.csv")
measurements_porespy

### Combine two dataframes

The two dataframes are now combined on their index.

# Combine dataframes
measurements = pd.concat([measurements_simpleitk, measurements_porespy], axis = 1)
measurements

measurements.to_csv(path + "02-measurements-concatenated.csv", index = False)
