==> quapos-lm/02-feature-extraction/01-wt/03-filter-relevant-features.ipynb <==
## Filter relevant features

Features were extracted and the dataframes [combined](02-combine-dataframes.ipynb). Not all features extracted from simpleitk are usefull for the downstream process. Relevant features will be filtered and kept for subsequent analysis.

import numpy as np
import pandas as pd

# Define path were features are stored
path = "../../measurements/wt-postnatal-development/"

# Load the measurements
measurements = pd.read_csv(path + "02-measurements-concatenated.csv")
measurements

# Show the column names
measurements.keys()

### Filter relevant features

Apply a filter to keep relevant features.

# Filter measurements
measurements_filtered = measurements[['label', 'maximum', 'mean', 'median', 'minimum', 'sigma', 'sum',
       'variance', 'bbox_3', 'bbox_4', 'bbox_5', 'elongation',
       'feret_diameter', 'flatness', 'equivalent_ellipsoid_diameter_0', 'equivalent_ellipsoid_diameter_1',
       'equivalent_ellipsoid_diameter_2', 'equivalent_spherical_perimeter',
       'equivalent_spherical_radius', 'volume', 'surface_area', 'bbox_volume', 'convex_volume', 'sphericity',
       'solidity', 'perimeter_2d', 'major_axis_length_2d',
       'minor_axis_length_2d', 'age', 'biol_repl', 'image_id']]
measurements_filtered

measurements_filtered.to_csv(path + "03-relevant-features.csv", index = False)
