==> quapos-lm/02-feature-extraction/01-wt/05-a-remove-inf-values.ipynb <==
## Data filtering - inf values

The feature extraction with porespy added under certain cases inf values for some 3D measurements (sphericity or convex volume). It is expected, that the classifier still segments smaller particles in 2D. When porespy tries to extract a 3D feature from a 2D label the inf value is returned. 

Since these inf values could harm the downstream analysis process, they will be removed here.

import numpy as np
import pandas as pd

# Define path were measurements are stored
path = "../../measurements/wt-postnatal-development/"

# Load the measurements
measurements = pd.read_csv(path + "04-measurements-rescaled.csv")
measurements

### Identify inf values

Check if there are inf values in the dataframe and how many rows contain inf values.

# Check for inf values
np.isinf(measurements.values).any()

# Check how many rows contain inf values
(measurements.isin([np.inf, -np.inf])).any(axis=1).sum()

### Remove inf values from the dataframe

# Remove inf values
measurements_without_inf = measurements[~measurements.isin([np.inf, -np.inf]).any(axis=1)]
measurements_without_inf

# Any infs left?
np.isinf(measurements_without_inf.values).any()

measurements_without_inf.to_csv(path + "05-a-measurements-inf-reduced.csv", index = False)
