==> quapos-lm/02-feature-extraction/01-wt/05-b-filter-5-and-95-percentiles.ipynb <==
## Filter dataset based on 5- and 95 percentiles

The classifier still predicts labels which correspond to several outer segments or which are too small to be considered a outer segment at all.

Here a filter of the 5- and 95 percentiles are applied based on the rescaled volume of the outer segments.

import numpy as np
import pandas as pd

# Define a path were the measurements are stored
path = "../../measurements/wt-postnatal-development/"

# Load the measurements
measurements = pd.read_csv(path + "05-a-measurements-inf-reduced.csv")
measurements

# Show how many OS are in the dataset
print("Number of OS in dataset before filter: ", len(measurements))

### Calculate 5-, and 95-percentile

From the dataset the 5-, and 95- percentiles are calculated to filter the dataset.

# Calculate five percentile of the volume
five_percentile = measurements["volume_rescaled"].quantile(0.05)
five_percentile

# Calculate the 95 percentile of the volume
ninety_five_percentile = measurements["volume_rescaled"].quantile(0.95)
ninety_five_percentile

### Filter for 5 percentile

A filter for the 5-percentile of the volume will be applied.

# Determine number of OS with volume smaller/bigger than the 5 percentile
print("Number of OS in dataset smaller than 5 percentile: ", len(measurements[measurements["volume_rescaled"] <= five_percentile]))
print("Number of OS in dataset bigger than 5 percentile: ", len(measurements[measurements["volume_rescaled"] > five_percentile]))

Apply the filter for the 5-percentile.

# Filter for 5 percentile of the volume
filtered_5_percentile_measurements = measurements[measurements["volume_rescaled"] > five_percentile]
filtered_5_percentile_measurements

### Filter for 95-percentile

Analog the filter for the 95-percentile will be applied.

# Determine number of OS with volume smaller/bigger than 95 percentile
print("Number of OS in dataset smaller than 95 percentile: ", len(filtered_5_percentile_measurements[filtered_5_percentile_measurements["volume_rescaled"] < ninety_five_percentile]))
print("Number of OS in dataset bigger than 95 percentile: ", len(filtered_5_percentile_measurements[filtered_5_percentile_measurements["volume_rescaled"] >= ninety_five_percentile]))

# Filter for 95-percentile
filtered_measurements = filtered_5_percentile_measurements[filtered_5_percentile_measurements["volume_rescaled"] < ninety_five_percentile]
filtered_measurements

filtered_measurements.to_csv(path + "05-b-measurements-filtered.csv", index=False)
