==> quapos-lm/02-feature-extraction/01-wt/06-average-values.ipynb <==
## Compute average values for statistics

The dataset was [rescaled](04-rescale-size-features.ipynb), and [inf values](05-a-remove-inf-values.ipynb) as well as the [5-, and 95-percentiles](05-b-filter-5-and-95-percentiles.ipynb) filtered.

For the subsequent statistics, the average values of the features should be plotted. Here a dataframe will be computed which contains average values from the filtered dataframe.

For the postnatal development 3 technical replicates per biological replicate were acquire. The following strategy is applied. Calculate the mean of every image/technical replicate, then compute the mean out of the means of the technical replicates.

import numpy as np
import pandas as pd

# Define path were measurements are stored
path = "../../measurements/wt-postnatal-development/"

# Load the measurements
measurements = pd.read_csv(path + "05-b-measurements-filtered.csv")
measurements

### Average value per image

The average values in every image are calculated.

# Calculate the means per image id
measurements_mean_image_id = measurements.groupby("image_id", as_index=False).mean()
measurements_mean_image_id

### Average per biological replicate

Now this dataframe is used to calculate the mean per biological replicate.

# Calculate mean in biological replicate
measurements_mean_biol_repl = measurements_mean_image_id.groupby("biol_repl", as_index=False).mean()
measurements_mean_biol_repl

### Filter columns: image_id and label

The 2 columns `image_id`, and `label` contain misleading values now. The average image_id per biological replicate is redundant. For the label instead of the mean number of label per image, the max number of labels per image should be determined. This will be done in the next notebook.

# Filter image_id and label
measurements_filtered = measurements_mean_biol_repl.drop(labels=["image_id", "label"], axis=1)
measurements_filtered

measurements_filtered.to_csv(path + "06-measurements-average.csv", index=False)
