==> quapos-lm/02-feature-extraction/01-wt/07-add-number-of-labels.ipynb <==
## Calculate max number of labels per image

In the [previous notebook](06-average-values.ipynb) average values of the features (per image and per biological replicate) were calculated for creating plots and statistics subsequently. Due to that the number of labels were misleading (should be maximum per image instead of average per image). 

This notebook computes the maximum number of labels from the filtered dataset and adds them to the average dataframe.

import numpy as np
import pandas as pd

# Define path of measurements
path = "../../measurements/wt-postnatal-development/"

# Load the filtered dataset
measurements = pd.read_csv(path + "05-b-measurements-filtered.csv")
measurements

### Determine max number of labels per image

Now the max number of labels per image are determined with the `df.groupby()` function. The argument `.size()` will be used instead of `.max()` since certain rows were filtered from the dataframe. This action requires to rename the resulting column from "size" into "label" again.

# Compute max number of labels
labels_per_image = measurements.groupby("image_id", as_index=False)["label"].size()

# Rename resulting column
labels_per_image.rename(columns={"size": "label"}, inplace=True)
labels_per_image

### Determine the average per biological replicate

The max number of label per image must be averaged per biological replicate. Hence, a dataframe containing the averaged biological replicate per image will be computed and merged to the labels per image dataframe.

# Compute average biological replicate per image
biol_repl = measurements.groupby("image_id", as_index=False)["biol_repl"].mean()

# Merge labels per image and biological replicate
labels_per_image = pd.merge(labels_per_image, biol_repl, how="inner")
labels_per_image

### Compute mean number of labels per biological replicate

Now that the dataframe is prepared the mean number of labels per biological replicate will be computed.

# Compute mean number of labels per biological replicate
labels_mean_per_biol_repl = labels_per_image.groupby("biol_repl", as_index=False)["label"].mean()
labels_mean_per_biol_repl

### Add mean number of labels to average dataset

Finally, the dataset containing the mean number of labels per biological replicate will be added to the dataframe containing average values.

# Load dataframe with means
measurements_means = pd.read_csv(path + "06-measurements-average.csv")

# Add mean number of labels to dataframe
measurements = pd.merge(labels_mean_per_biol_repl, measurements_means, how="inner")
measurements

measurements.to_csv(path + "07-os-numbers-added.csv", index=False)
