==> quapos-lm/02-feature-extraction/01-wt/08-add-summed-volume.ipynb <==
## Summed OS volume

One way to circumvent that several outer segments are combined in one label (without filtering) is calculating the summed OS volume per image and afterwards the average per biological replicate.

import numpy as np
import pandas as pd

# Define a file path with the measurements
path = "../../measurements/wt-postnatal-development/"

# Open unfiltered dataset
measurements = pd.read_csv(path + "05-a-measurements-inf-reduced.csv")
measurements

### Determine summed volume per image

Now from the unfiltered dataset the summed volume per image is computed.

# Determine summe volume
summed_volume = measurements.groupby("image_id", as_index=False)["volume_rescaled"].sum()
summed_volume

### Determine mean summed volume per biological replicate

Again, the mean of the biological replicate per image id will be calculated and added to the summed volume. Afterwards the mean of the summed volume per biological replicate will be computed.

# Determine mean of biological replicate
biol_repl = measurements.groupby("image_id", as_index=False)["biol_repl"].mean()

# Add biological replicate to dataframe
summed_volume = pd.merge(summed_volume, biol_repl, how="inner")

# Determine the mean of the summed volume per biological replicate
summed_volume_mean = summed_volume.groupby("biol_repl", as_index=False)["volume_rescaled"].mean()
summed_volume_mean.rename(columns={"volume_rescaled": "summed_volume_mean"}, inplace=True)
summed_volume_mean

### Add mean summed volume to average dataframe

Finally, the mean summed volume will be added to the average dataframe.

# Load dataframe containing mean values
measurements_means = pd.read_csv(path + "07-os-numbers-added.csv")

# Add summed mean volume per biological replicate
measurements = pd.merge(measurements_means, summed_volume_mean, how="inner")
measurements

measurements.to_csv(path + "08-measurements-processed.csv", index=False)
